Port changes to native Azure project
