+++ MasterDetail-Native-Azure/MasterDetail.UWP/MasterDetail.UWP.csproj
@@ -122,6 +122,7 @@
   </ItemGroup>
   <ItemGroup>
     <Content Include="Assets\profile_generic.png" />
+    <Content Include="Assets\xamarin_logo.png" />
     <Content Include="Properties\Default.rd.xml" />
     <Content Include="Assets\LockScreenLogo.scale-200.png" />
     <Content Include="Assets\SplashScreen.scale-200.png" />
