+++ MasterDetail-Native-Azure/MasterDetail.UWP/Views/MainPivot.xaml
@@ -7,43 +7,39 @@
     xmlns:viewmodels="using:MasterDetail.ViewModel"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d">
-    <Page.DataContext>
-        <viewmodels:ItemsViewModel/>
-    </Page.DataContext>
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="10*"/>
-            <RowDefinition Height="90*"/>
-        </Grid.RowDefinitions>
-        <Border  Grid.Row="0" Background="#0078d7" />
-        <TextBlock x:Name="txtTitle"
-                   VerticalAlignment="Center"
-                   Grid.Row="0"
-                   FontSize="36"
-                   Text="AppName"
-                   Foreground="White"/>
-        <Pivot x:Name="rootPivot" Grid.Row="1">
-            <PivotItem Header="Browse" DataContext="{Binding BrowseItemsViewModel}">
-                <ListView x:Name="gvItems" ItemsSource="{Binding Items,Mode=OneWay}" SelectionMode="Single" IsItemClickEnabled="True">
-                    <ListView.ItemTemplate>
-                        <DataTemplate x:DataType="data:Item">
-                            <StackPanel>
-                                <TextBlock x:Name="txtText" Text="{Binding Text}"/>
-                                <TextBlock Text="{Binding Description}"/>
-                            </StackPanel>
-                        </DataTemplate>
-                    </ListView.ItemTemplate>
-                </ListView>
-            </PivotItem>
-            <PivotItem Header="About">
-                <StackPanel>
-                    <Border Background="#96d1ff" Padding="25">
-                        <Image Source="/Assets/xamarin_logo.png" Grid.Row="0" Height="70"/>
-                    </Border>
+    mc:Ignorable="d"
+    DataContext="{Binding RelativeSource={RelativeSource Self}}"
+    Background="{StaticResource PrimaryColor}">
+
+    <Page.Resources>
+        <Style TargetType="PivotItem">
+            <Setter Property="Background" Value="White"/>
+            <Setter Property="Margin" Value="0"/>
+        </Style>
+
+        <DataTemplate x:Key="GvItemTemplate" x:DataType="data:Item">
+            <StackPanel Margin="0,10">
+                <TextBlock x:Name="txtText" Text="{Binding Text}" FontSize="16" FontWeight="Bold" Margin="0,5"/>
+                <TextBlock Text="{Binding Description}" FontSize="14"/>
+            </StackPanel>
+        </DataTemplate>
+    </Page.Resources>
+
+    <Pivot x:Name="rootPivot" Title="AppName" SelectionChanged="PivotItemChanged">
+        <PivotItem Header="{Binding Title}" DataContext="{Binding BrowseViewModel}">
+            <ListView ItemsSource="{Binding Items,Mode=OneWay}" ItemTemplate="{StaticResource GvItemTemplate}"
+                      ItemClick="GvItems_ItemClick" SelectionMode="Single" IsItemClickEnabled="True">
+            </ListView>
+        </PivotItem>
+        <PivotItem Header="{Binding Title}" DataContext="{Binding AboutViewModel}">
+            <StackPanel>
+                <Border Background="{StaticResource AccentColor}" Padding="25">
+                    <Image Source="/Assets/xamarin_logo.png" Height="70"/>
+                </Border>
+                <StackPanel Padding="10">
                     <TextBlock Margin="0,15" FontSize="20">
                             <Bold>AppName&#160;</Bold>
-                            <Span Foreground="#999999">1.0</Span>
+                            <Span Foreground="{StaticResource LightTextColor}">1.0</Span>
                     </TextBlock>
                     <TextBlock TextWrapping="WrapWholeWords">
                             This app is written in C# and native APIs using the
@@ -51,16 +47,15 @@
                             <Bold>iOS, Android, and Windows</Bold> versions.
                     </TextBlock>
                     <Button Content="Learn More" Command="{Binding OpenWebCommand}" Margin="0,20" 
-                            Background="#2196F3" Foreground="White" HorizontalAlignment="Stretch"></Button>
+                            Background="{StaticResource PrimaryColor}" Foreground="White" HorizontalAlignment="Stretch"></Button>
                 </StackPanel>
-            </PivotItem>
-        </Pivot>
-    </Grid>
-    <Page.BottomAppBar>
-        <AppBar IsOpen="False" IsSticky="False">
-            <StackPanel Orientation="Horizontal">
-                <AppBarButton Name="Show" Icon="Add" Label="Add" Click="Show_Click"/>
             </StackPanel>
-        </AppBar>
+        </PivotItem>
+    </Pivot>
+
+    <Page.BottomAppBar>
+        <CommandBar Name="Toolbar" ClosedDisplayMode="Compact">
+            <AppBarButton Name="AddItemBtn" Icon="Add" Label="Add" Click="AddItem_Click"/>
+        </CommandBar>
     </Page.BottomAppBar>
 </Page>
