+++ MasterDetail-Native/MasterDetail.UWP/Views/BrowseItemDetail.xaml.cs
@@ -1,17 +1,12 @@
-﻿using Windows.UI.Xaml;
+﻿using MasterDetail.Model;
+using MasterDetail.ViewModel;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Navigation;
-using MasterDetail.Model;
-using MasterDetail.ViewModel;
 
 namespace MasterDetail.UWP.Views
 {
-    /// <summary>
-    /// An empty page that can be used on its own or navigated to within a Frame.
-    /// </summary>
     public sealed partial class BrowseItemDetail : Page
     {
-
         public ItemDetailViewModel ViewModel { get; set; }
         int quantityCount = 0;
         public BrowseItemDetail()
@@ -22,10 +17,6 @@ namespace MasterDetail.UWP.Views
             DataContext = new ItemDetailViewModel();
         }
 
-        public void Show_Click(object sender, RoutedEventArgs e)
-        {
-            System.Diagnostics.Debug.WriteLine("yay");
-        }
         protected override void OnNavigatedTo(NavigationEventArgs e)
         {
             ViewModel.Item = (Item)e.Parameter;
@@ -34,5 +25,4 @@ namespace MasterDetail.UWP.Views
             txtDesc.Text = ViewModel.Item.Description;
         }
     }
-
 }
